Import GetStatus/SetStatus from the info component, where they were moved in a3ec4d4.
